rewrote how the doAction() function works in each controller and Walleye.
Now you pass the action and data as parameters
of the doAction() function instead using Walleye
to retrieve that information from its properties.

Also, rewrote Walleye to be more secure and easier
to understand.

Removed storage of the currently logged in user/setting
the currently logged in user from Walleye to
mUser